### 2017-03-30 #218 @whedon
[PRE REVIEW]: Noisyopt
**Submitting author:** @andim (<a href="http://orcid.org/0000-0002-6643-7622">Andreas Mayer</a>)
**Repository:** <a href="https://github.com/andim/noisyopt" target ="_blank">https://github.com/andim/noisyopt</a>
**Version:** v0.2
