+++ README.md
@@ -1,11 +1,13 @@
 [![License](https://img.shields.io/pypi/l/noisyopt.svg)](https://github.com/andim/noisyopt/blob/master/LICENSE)
-![Status](https://img.shields.io/pypi/status/noisyopt.svg)
 [![Latest release](https://img.shields.io/pypi/v/noisyopt.svg)](https://pypi.python.org/pypi/noisyopt)
+![Py2.7](https://img.shields.io/badge/python-2.7-blue.svg)
+![Py3.5](https://img.shields.io/badge/python-3.5-blue.svg)
+[![DOI](https://zenodo.org/badge/54976198.svg)](https://zenodo.org/badge/latestdoi/54976198)
+
+![Status](https://img.shields.io/pypi/status/noisyopt.svg)
 [![Build Status](https://travis-ci.org/andim/noisyopt.svg?branch=master)](https://travis-ci.org/andim/noisyopt)
 [![Documentation Status](https://readthedocs.org/projects/noisyopt/badge/?version=latest)](https://noisyopt.readthedocs.io/en/latest/?badge=latest)
 [![Coverage Status](https://coveralls.io/repos/github/andim/noisyopt/badge.svg?branch=master)](https://coveralls.io/github/andim/noisyopt?branch=master)
-[![DOI](https://zenodo.org/badge/54976198.svg)](https://zenodo.org/badge/latestdoi/54976198)
-
 
 
 # noisyopt
