### 2017-04-05 #218 @arokem
:hand: I can edit this one

@andim : do you have suggestions for reviewers? There is a list of potential reviewers here: https://github.com/openjournals/joss/blob/master/docs/reviewers.csv

But if there is someone else not on this list who would be an appropriate reviewer, please free to suggest them.
