### 2017-04-05 #218 @arfon
> @whedon set @arokem as editor

Sorry, there's a small inconsistency in language. `assign` for users, `set` for the DOI. I should fix that...
