### 2017-04-07 #218 @andim
Outside of the list @GaelVaroquaux might be a good choice. He wrote the excellent scipy-lecture notes on optimization (http://www.scipy-lectures.org/advanced/mathematical_optimization/index.html) so knows well what else exists within the Python ecosystem.

If you have other ideas @arokem feel free to assign someone else.
