### 2017-04-07 #218 @arokem
@moorepants : would you be interested in reviewing this software for JOSS?
