### 2017-04-16 #218 @arokem
OK - no response there. Let's try someone else.

@jordigh: would you be willing to review this submission for JOSS?

Thanks!
