### 2017-05-08 #218 @arokem
No response there.

@nirum: would you like to review this for JOSS?
