### 2017-05-11 andim/noisyopt#4 @nirum
JOSS review: installing dependencies via pip
Hi,

I'm trying to install `noisyopt` in a clean conda environment running python 3.5. Running `pip install noisyopt` does not install the dependencies (numpy, scipy). I see that you do include a `requires` keyword argument in your `setup.py` file, does this need to be `install_requires`? (as in https://packaging.python.org/requirements/).

Also, not necessary if you don't want to, but I think it would be good to include a list of dependences somewhere in the README.
