### 2017-05-11 andim/noisyopt#4 @nirum
(edit: just noticed the dependencies are listed in the Installation section in the README)
