### 2017-05-11 7c0c40e
setup install_requires instead of requires
