+++ setup.py
@@ -23,7 +23,7 @@ opts = dict(name=NAME,
             version=VERSION,
             packages=PACKAGES,
             package_data=PACKAGE_DATA,
-            requires=REQUIRES)
+            install_requires=REQUIRES)
 
 
 if __name__ == '__main__':
