### 2017-05-11 andim/noisyopt#4 @andim
Thanks for noticing this! Fixed this in 7c0c40ee4d9d229e9514f4e4e33695d409ecfed5
