### 2017-05-11 andim/noisyopt#4 @nirum
Thanks! Mind pushing an update to pypi? you could just bump the version to 0.2.1
