+++ examples/basic.py
@@ -1,0 +1,13 @@
+# This is a minimal usage example
+# see also http://noisyopt.readthedocs.io/en/latest/tutorial.html
+
+import numpy as np
+from noisyopt import minimizeCompass
+
+def obj(x):
+    return (x**2).sum() + 0.1*np.random.randn(1)
+
+bounds = [[-3.0, 3.0], [0.5, 5.0]]
+x0 = np.array([-2.0, 2.0])
+res = minimizeCompass(obj, bounds=bounds, x0=x0, deltatol=0.1, paired=False)
+print(res)
