+++ doc/tutorial.rst
@@ -32,7 +32,7 @@ First we need to import the ``minimizeCompass`` function from the `noisyopt` pac
 Then we need to define the objective of the function::
 
   >>> def obj(x):
-    ...     return (x**2).sum() + np.random.randn(1)
+    ...     return (x**2).sum() + 0.1*np.random.randn(1)
 
 We now define the domain of the problem using bound constraints::
 
@@ -42,9 +42,9 @@ And we define the initial guess::
 
   >>> x0 = np.array([-2.0, 2.0])
                
-The pattern search based optimization algorithm is called using the ``minimizeCompass`` function. The `minimizeCompass` functions accepts the problem objective ``obj`` and block constraints::
+The pattern search based optimization algorithm is called using the ``minimizeCompass`` function. The `minimizeCompass` functions accepts the problem objective ``obj`` and bound constraints::
 
-  >>> res = minimizeCompass(obj, bounds=bounds, x0=x0, deltatol=0.1)
+  >>> res = minimizeCompass(obj, bounds=bounds, x0=x0, deltatol=0.1, paired=False)
 
 It is possible to further customize the algorithm using the parameters of
 the ``minimizeCompass`` function (see :py:func:`noisyopt.minimizeCompass`). Alternatively we can also try a different algorithm implementing a simultaneous perturbation stochastic approximation algorithm (see :py:func:`noisyopt.minimizeSPSA`).
@@ -54,15 +54,15 @@ other the optimal point, ``res.x``, and the value of the objective at the
 optimum, ``res.fun``::
 
   >>> print(res)
-       funse: 0.0089061854063998726
-     success: True
-        free: array([False, False], dtype=bool)
-        nfev: 1320
-         fun: 0.23298649773605173
-           x: array([-0.05,  0.5 ])
-     message: 'convergence within deltatol'
-         nit: 10
-
+      free: array([False, False], dtype=bool)
+       fun: 0.25068227287617101
+     funse: 0.0022450327079771111
+   message: 'convergence within deltatol'
+      nfev: 9510
+       nit: 9
+   success: True
+         x: array([ 0. ,  0.5])
+  
 Further examples
 ----------------
 
