### 2017-05-11 8a965ad
readme updated with quick start section and links to other optimization packages
