### 2017-05-11 b2f7f4d
bumped version to 0.2.1
