### 2017-05-15 ee08570
travis added python 3.6
