+++ .travis.yml
@@ -7,6 +7,7 @@ deploy:
   user: andreas.mayer
   password:
     secure: Muws+2oiSf4TM40LE+uhPq1PcX8jiN3TsNEFuqaccIWQCBId7RHCj620TWqDoBQNSb9BIlokVCx+nEewmQ4yzwUpQwx8FguRAJnkTNcwNJZquF4kmVJ4iyAxH/EDn2MikZU3RqA8NVqqda+eZBV9L1YbrZ5z+g9RhDr8U4PmqIlv6Cnu3lGmWt8pcL91RMkMJYDw6hKl1XS5Ghe/Ukqa+wmYwPwteCZFx9xsgSmVRDKgS0ycXdtDuoc0ewO4UWgDooMB7DKcHCWhIpYRDRxCJkGHQ4I5L6SrIF4fPtiz2pY9nNDoOHLdLfUSIjFA5X5AtIzdNbtbOIfFn9h11y1eaHXoqWibbfG49N4HR9U5sdHy03IgBE1SVejP0UF0z1F4WJh6cf3QC8XYWcZgYtNGFUUM/xsel6HCZ9xDjPx2T2NyeeCGEnMkkafBhwZBYgm+KhTy+404pDaqZYknN89jyYNJ9K72zyM4oCV643ubCU3yLfiZml7oWc7/Ui9eiGXDky2Tqv6gF420LQZTBKlcdmTcLj6trcC/SZhKU3rIqaJTaRNMHmfAF2YTBlo3OExthZyN1PKWeqjmt9wDXhQHB+S/QCnGnJMj5VkQLkvIRWS5HpB74MdEL4IjDfEyf/XnPmdXzMWqon+jjBoM7vOYA3BL6zr4C3aq1MLIdXiO6OI=
+  distributions: "sdist bdist_wheel"
   on:
     tags: true
 
