### 2017-05-15 9412aec
paper updated with feedback from @nirum
