+++ paper.bib
@@ -1,3 +1,13 @@
+@article{adaptpattern,
+author = {Mayer, Andreas and Mora, Thierry and Rivoire, Olivier and Walczak, Aleksandra M},
+title = {{Diversity of immune strategies explained by adaptation to pathogen statistics}},
+journal = {Proceedings of the National Academy of Sciences},
+number = {31},
+pages = {8630--8635},
+volume = {113},
+year = {2016}
+}
+
 @Misc{scipy,
   author = {Eric Jones and Travis Oliphant and Pearu Peterson and others},
   title  = {{SciPy}: Open source scientific tools for {Python}},
@@ -5,4 +15,16 @@
   url    = "http://www.scipy.org/",
   note   = {[Online; accessed 2017/03/30]}
 }
+
+
+@article{spsa,
+author = {Spall, James C},
+title = {{Implementation of the Simulteaneous Perturbation Algorithm for Stochastic Optimization}},
+journal = {IEEE Transactions on Aerospace and Electronic Systems},
+year = {1998},
+number = {3},
+volume = {34},
+pages = {817 -- 823},
+}
+
 
