### 2017-05-15 8657d4a
fixed minimizeCompass docstring escape chars
