### 2017-05-15 798dd7c
doc: algos do not require explicit derivatives
