### 2017-05-15 2e47cd5
added pending deprecation warning to address issue #11
