### 2017-05-15 67ad9e1
added version classifiers for pypi
