+++ noisyopt/version.py
@@ -22,6 +22,11 @@ CLASSIFIERS = ["Development Status :: 5 - Production/Stable",
                "License :: OSI Approved :: MIT License",
                "Operating System :: OS Independent",
                "Programming Language :: Python",
+               "Programming Language :: Python :: 2",
+               "Programming Language :: Python :: 2.7",              
+               "Programming Language :: Python :: 3",              
+               "Programming Language :: Python :: 3.5",              
+               "Programming Language :: Python :: 3.6",              
                "Topic :: Scientific/Engineering"]
 
 # Description should be a one-liner:
