### 2017-05-15 4193cf7
tutorial: mention that you get constrained optimum
