### 2017-05-15 1694814
added more unit tests
