### 2017-05-15 28d17fd
README updated with joss badge and python versions from pypi
