### 2017-05-23 d3ce9bc
better documentation of spsa algorithm
