### 2017-05-23 29cbe6d
added spsa example to tutorial
