### 2017-05-23 0eba0e0
enhanced minimizeCompass docstring
