### 2017-05-23 9599387
added minimizeCompass termination criterion to docstring
