### 2017-05-23 2ee8f9c
added a few more tests
