### 2017-05-23 #258 @nirum
The author(s) have addressed all of my comments and closed the issues, which are linked below.

https://github.com/andim/noisyopt/issues/4
