### 2017-05-23 #258 @nirum
@arfon @arokem what's the next step here? It looks like @andim has already created a DOI for the package:

[![DOI](https://zenodo.org/badge/54976198.svg)](https://zenodo.org/badge/latestdoi/54976198)
